edureach/ai-service/requirements.txt
fastapi
uvicorn
librosa
scikit-learn
transformers
python-dotenv
numpy
torch
openai
groq
httpx
python-multipart
aiohttp
